requirements-test.txt
# Testing dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.11.1
pytest-xdist>=3.3.1
pytest-timeout>=2.1.0
pytest-env>=0.8.2

# HTTP testing
httpx>=0.24.1
requests-mock>=1.11.0

# Factories & Fixtures
factory-boy>=3.3.0
faker>=19.2.0

# Database testing
sqlalchemy-utils>=0.41.1

# Code quality
black>=23.7.0
flake8>=6.1.0
mypy>=1.4.1
isort>=5.12.0
bandit>=1.7.5

# E2E Testing
playwright>=1.40.0
pytest-playwright>=0.4.3
